// src/lib/types.ts
export type ID = string;

export interface BaseEntity {
  id: ID;
  organization_id: ID;
  created_at: string;
  updated_at: string;
}
// ...
export type Priority = "Low" | "Medium" | "High" | "Urgent";
// ...
export type TaskStatus =
  | "Backlog"
  | "To Do"
  | "In Progress"
  | "Review"
  | "Completed"
  | "Blocked";

export interface ChecklistItem {
  id: ID;
  title: string;
  done: boolean;
}

export interface Task extends BaseEntity {
  code: string;
  title: string;
  description: string;
  project_id: ID;
  assignee_ids: ID[];
  team_id: ID | null;
  reporter_id: ID;
  follower_ids: ID[];
  priority: Priority;
  status: TaskStatus;
  start_date: string;
  due_date: string;
  progress: number;
  checklist: ChecklistItem[];
  parent_task_id: ID | null;
  dependency_ids: ID[];
  labels: string[];
  attachment_ids: ID[];
  milestone_id: ID | null;
}